refactor: Phase 3 slice 3a — App controller + IPC router pattern (#84)
Introduces the App/Core lifecycle controller and the per-domain IPC router
pattern that the rest of Phase 3 (#102) will fill in. Behavior-identical,
zero IPC surface change.

* New src/main/app/core.ts owns command-line switches, anime-video://
  scheme registration, EPIPE silencing, app.whenReady() gating, and the
  window-all-closed / before-quit handlers — all previously top-of-file
  side effects in src/main/index.ts. The whenReady body becomes a
  bootstrap() function passed to App.start({ onReady, onBeforeQuit }).

* New src/main/ipc/index.ts exposes registerIpcRouters(deps) and the
  AppDeps shape; subsequent slices extend AppDeps and add more routers.
  Legacy registerIpcHandlers() still owns the remaining 113 handlers,
  invoking registerIpcRouters first.

* New src/main/ipc/app.ipc.ts is the first per-domain router. It owns
  the four app/update handlers (CHANNELS.APP_VERSION, UPDATE_CHECK,
  UPDATE_DOWNLOAD, UPDATE_INSTALL), the five autoUpdater event
  subscriptions broadcasting EVENT_CHANNELS.UPDATE_STATUS, and the 24h
  auto-check-on-launch trigger gated by store.get('lastUpdateCheck').

* test/ipc-channels.test.ts widens its source set to src/main/ipc/**
  so the drift-guard (no dead contract constants) sees channels
  referenced from routers, not just src/main/index.ts.

* DESIGN.md adds entries for the new app/ and ipc/ modules.

Refs #102 (Phase 3 tracking issue), #84 (refactor epic).
